Refactor : Removes login module because it takes time to handle module configs by my internet and vpn state so I removed it temporary and moved back every thing to app
