+++ app/build.gradle
@@ -6,14 +6,14 @@ apply plugin: 'kotlin-parcelize'
 
 android {
     compileSdkVersion 34
-    buildToolsVersion "29.0.1"
+    buildToolsVersion "34.0.0"
     repositories {
         google()
         mavenCentral()
     }
     defaultConfig {
         applicationId "com.jabama.challenge.github"
-        minSdkVersion 17
+        minSdkVersion 21
         targetSdkVersion 29
         versionCode 1
         versionName "1.0"
@@ -59,7 +59,6 @@ dependencies {
     implementation 'androidx.preference:preference-ktx:1.1.0'
     implementation 'com.squareup.retrofit2:converter-gson:2.4.0'
     implementation 'com.jakewharton.retrofit:retrofit2-kotlin-coroutines-adapter:0.9.2'
-    implementation project(':login')
 
     testImplementation 'junit:junit:4.12'
     androidTestImplementation 'androidx.test:runner:1.1.1'
