+++ app/src/main/java/com/jabama/challenge/data/repository/AccessTokenDataSourceImpl.kt
@@ -1,0 +1,11 @@
+package com.jabama.challenge.data.repository
+
+import com.jabama.challenge.data.api.AccessTokenService
+import com.jabama.challenge.data.model.RequestAccessToken
+import com.jabama.challenge.domain.AccessTokenDataSource
+
+const val TOKEN = "TOKEN"
+class AccessTokenDataSourceImpl(private val accessTokenService: AccessTokenService) :
+    AccessTokenDataSource {
+    override fun accessToken(requestAccessToken: RequestAccessToken) = accessTokenService.accessToken(requestAccessToken)
+}
