+++ app/src/main/java/com/jabama/challenge/domain/AccessTokenDataSource.kt
@@ -1,0 +1,7 @@
+package com.jabama.challenge.domain
+
+import kotlinx.coroutines.Deferred
+
+interface AccessTokenDataSource {
+    fun accessToken(requestAccessToken: com.jabama.challenge.data.model.RequestAccessToken): Deferred<com.jabama.challenge.data.model.ResponseAccessToken>
+}
