--- login/src/main/res/values/themes.xml
@@ -1,5 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-
-    <style name="Theme.Github" parent="android:Theme.Material.Light.NoActionBar" />
-</resources>
