+++ settings.gradle
@@ -1,3 +1,2 @@
 include ':app'
 rootProject.name='Github'
-include ':login'
